+++ src/query-performance-monitoring/common-parameters/common_parameters.go
@@ -1,6 +1,8 @@
 package commonparameters
 
 import (
+	"strings"
+
 	"github.com/newrelic/infra-integrations-sdk/v3/log"
 	"github.com/newrelic/nri-postgresql/src/args"
 )
@@ -21,6 +23,7 @@ type CommonParameters struct {
 	QueryMonitoringResponseTimeThreshold int
 	Host                                 string
 	Port                                 string
+	IsRds                                bool
 }
 
 func SetCommonParameters(args args.ArgumentList, version uint64, databases string) *CommonParameters {
@@ -31,7 +34,12 @@ func SetCommonParameters(args args.ArgumentList, version uint64, databases strin
 		QueryMonitoringResponseTimeThreshold: validateAndGetQueryMonitoringResponseTimeThreshold(args),
 		Host:                                 args.Hostname,
 		Port:                                 args.Port,
+		IsRds:                                isRdsHost(args.Hostname),
 	}
+}
+
+func isRdsHost(hostname string) bool {
+	return strings.Contains(hostname, ".rds.amazonaws.com")
 }
 
 func validateAndGetQueryMonitoringResponseTimeThreshold(args args.ArgumentList) int {