+++ src/query-performance-monitoring/common-utils/query_fetch_helpers.go
@@ -4,7 +4,7 @@ import (
 	"github.com/newrelic/nri-postgresql/src/query-performance-monitoring/queries"
 )
 
-func FetchVersionSpecificSlowQueries(version uint64) (string, error) {
+func FetchVersionSpecificSlowQuery(version uint64) (string, error) {
 	switch {
 	case version == PostgresVersion12:
 		return queries.SlowQueriesForV12, nil
@@ -15,11 +15,13 @@ func FetchVersionSpecificSlowQueries(version uint64) (string, error) {
 	}
 }
 
-func FetchVersionSpecificBlockingQueries(version uint64) (string, error) {
+func FetchVersionSpecificBlockingQuery(version uint64, isRds bool) (string, error) {
 	switch {
 	case version == PostgresVersion12, version == PostgresVersion13:
 		return queries.BlockingQueriesForV12AndV13, nil
-	case version >= PostgresVersion14:
+	case version >= PostgresVersion14 && isRds:
+		return queries.RDSPostgresBlockingQueryForV14AndAbove, nil
+	case version >= PostgresVersion14 && !isRds:
 		return queries.BlockingQueriesForV14AndAbove, nil
 	default:
 		return "", ErrUnsupportedVersion
@@ -34,5 +36,13 @@ func FetchVersionSpecificIndividualQueries(version uint64) (string, error) {
 		return queries.IndividualQuerySearchV13AndAbove, nil
 	default:
 		return "", ErrUnsupportedVersion
+	}
+}
+
+func FetchSupportedWaitEventsQuery(isRDS bool) (string, error) {
+	if isRDS {
+		return queries.WaitEventsFromPgStatActivity, nil
+	} else {
+		return queries.WaitEvents, nil
 	}
 }