+++ src/query-performance-monitoring/queries/queries.go
@@ -73,7 +73,7 @@ const (
 	LIMIT
 		 %d; -- Limit the number of results`
 
-	// WaitEvents retrieves wait events and their statistics
+	// WaitEvents retrieves wait events and their statistics from pg_wait_sampling_history
 	WaitEvents = `WITH wait_history AS (
 		SELECT
 			wh.pid, -- Process ID
@@ -111,6 +111,42 @@ const (
 	ORDER BY total_wait_time_ms DESC -- Order by the total wait time in descending order
 	LIMIT %d; -- Limit the number of results`
 
+	// WaitEvents retrieves wait events and their statistics from pg_stat_activity
+	WaitEventsFromPgStatActivity = `WITH wait_history AS (
+        SELECT
+            sa.pid, -- Process ID
+            sa.wait_event_type AS event_type, -- Type of the wait event
+            sa.wait_event AS event, -- Wait event           
+            sa.backend_start AS duration, -- Timestamp of the wait event
+            pg_database.datname AS database_name, -- Name of the database
+            LEFT(ss.query, 4095) AS query_text, -- Query text truncated to 4095 characters
+            ss.queryid AS query_id -- Unique identifier for the query
+        FROM
+            pg_stat_activity sa
+        LEFT JOIN
+            pg_stat_statements ss ON  AnonymizeQueryText(sa.query) = AnonymizeQueryText(ss.query)
+        LEFT JOIN
+            pg_database ON pg_database.oid = sa.datid
+        WHERE pg_database.datname in (%s) -- List of database names
+    )
+    SELECT
+        event_type || ':' || event AS wait_event_name, -- Concatenated wait event name
+        CASE
+            WHEN event_type IN ('LWLock', 'Lock') THEN 'Locks' -- Wait category is Locks
+            WHEN event_type = 'IO' THEN 'Disk IO' -- Wait category is Disk IO
+            WHEN event_type = 'CPU' THEN 'CPU' -- Wait category is CPU
+            ELSE 'Other' -- Wait category is Other
+        END AS wait_category, -- Category of the wait event
+        to_char(NOW() AT TIME ZONE 'UTC', 'YYYY-MM-DD"T"HH24:MI:SS"Z"') AS collection_timestamp, -- Timestamp of data collection
+        query_id, -- Unique identifier for the query
+        query_text, -- Query text
+        database_name -- Name of the database
+    FROM wait_history
+    WHERE query_text NOT LIKE 'EXPLAIN (FORMAT JSON) %%' AND query_id IS NOT NULL AND event_type IS NOT NULL
+    GROUP BY event_type, event, query_id, query_text, database_name,duration
+    ORDER BY duration DESC -- Order by the total wait time in descending order
+    LIMIT %d;  -- Limit the number of results`
+
 	// BlockingQueriesForV14AndAbove retrieves information about blocking and blocked queries for PostgreSQL version 14 and above
 	BlockingQueriesForV14AndAbove = `SELECT 'newrelic' as newrelic, -- Common value to filter with like operator in slow query metrics
 		  blocked_activity.pid AS blocked_pid, -- Process ID of the blocked query
@@ -137,6 +173,38 @@ const (
 		  AND blocked_locks.pid <> blocking_locks.pid
 		JOIN pg_stat_activity AS blocking_activity ON blocking_locks.pid = blocking_activity.pid
 		JOIN pg_stat_statements AS blocking_statements ON blocking_activity.query_id = blocking_statements.queryid
+		WHERE NOT blocked_locks.granted
+		  AND blocked_activity.datname IN (%s) -- List of database names
+		  AND blocked_statements.query NOT LIKE 'EXPLAIN (FORMAT JSON) %%' -- Exclude EXPLAIN queries
+		  AND blocking_statements.query NOT LIKE 'EXPLAIN (FORMAT JSON) %%' -- Exclude EXPLAIN queries
+		ORDER BY blocked_activity.query_start ASC -- Order by the start time of the blocked query in ascending order
+		LIMIT %d; -- Limit the number of results`
+
+	RDSPostgresBlockingQueryForV14AndAbove = `SELECT 'newrelic' as newrelic, -- Common value to filter with like operator in slow query metrics
+		  blocked_activity.pid AS blocked_pid, -- Process ID of the blocked query
+		  LEFT(blocked_statements.query, 4095) AS blocked_query, -- Blocked query text truncated to 4095 characters
+		  blocked_statements.queryid AS blocked_query_id, -- Unique identifier for the blocked query
+		  blocked_activity.query_start AS blocked_query_start, -- Start time of the blocked query
+		  blocked_activity.datname AS database_name, -- Name of the database
+		  blocking_activity.pid AS blocking_pid, -- Process ID of the blocking query
+		  LEFT(blocking_statements.query, 4095) AS blocking_query, -- Blocking query text truncated to 4095 characters
+		  blocking_statements.queryid AS blocking_query_id, -- Unique identifier for the blocking query
+		  blocking_activity.query_start AS blocking_query_start -- Start time of the blocking query
+		FROM pg_stat_activity AS blocked_activity
+		JOIN pg_stat_statements AS blocked_statements ON AnonymizeQueryText(blocked_activity.query) = AnonymizeQueryText(blocked_statements.query)
+		JOIN pg_locks blocked_locks ON blocked_activity.pid = blocked_locks.pid
+		JOIN pg_locks blocking_locks ON blocked_locks.locktype = blocking_locks.locktype
+		  AND blocked_locks.database IS NOT DISTINCT FROM blocking_locks.database
+		  AND blocked_locks.relation IS NOT DISTINCT FROM blocking_locks.relation
+		  AND blocked_locks.page IS NOT DISTINCT FROM blocking_locks.page
+		  AND blocked_locks.tuple IS NOT DISTINCT FROM blocking_locks.tuple
+		  AND blocked_locks.transactionid IS NOT DISTINCT FROM blocking_locks.transactionid
+		  AND blocked_locks.classid IS NOT DISTINCT FROM blocking_locks.classid
+		  AND blocked_locks.objid IS NOT DISTINCT FROM blocking_locks.objid
+		  AND blocked_locks.objsubid IS NOT DISTINCT FROM blocking_locks.objsubid
+		  AND blocked_locks.pid <> blocking_locks.pid
+		JOIN pg_stat_activity AS blocking_activity ON blocking_locks.pid = blocking_activity.pid
+		JOIN pg_stat_statements AS blocking_statements ON AnonymizeQueryText(blocking_activity.query) = AnonymizeQueryText(blocking_statements.query)
 		WHERE NOT blocked_locks.granted
 		  AND blocked_activity.datname IN (%s) -- List of database names
 		  AND blocked_statements.query NOT LIKE 'EXPLAIN (FORMAT JSON) %%' -- Exclude EXPLAIN queries
@@ -193,6 +261,8 @@ const (
 		ORDER BY
 		 exec_time_ms DESC -- Order by average execution time in descending order
 		LIMIT %d; -- Limit the number of results`
+
+	IndividualQueryFromPgStat = "select query  from pg_stat_activity where query is not null and query !='';"
 
 	// IndividualQuerySearchV12 retrieves individual query statistics for PostgreSQL version 12
 	IndividualQuerySearchV12 = `SELECT 'newrelic' as newrelic, -- Common value to filter with like operator in slow query metrics