+++ src/query-performance-monitoring/validations/performance_metrics_validations_test.go
@@ -13,7 +13,7 @@ func TestCheckBlockingSessionMetricsFetchEligibilityExtensionNotRequired(t *test
 	conn, mock := connection.CreateMockSQL(t)
 	version := uint64(12)
 	enabledExtensions, _ := FetchAllExtensions(conn)
-	isExtensionEnabledTest, _ := CheckBlockingSessionMetricsFetchEligibility(enabledExtensions, version)
+	isExtensionEnabledTest := CheckBlockingSessionMetricsFetchEligibility(enabledExtensions, version)
 	assert.Equal(t, isExtensionEnabledTest, true)
 	assert.NoError(t, mock.ExpectationsWereMet())
 }
@@ -24,7 +24,7 @@ func TestCheckBlockingSessionMetricsFetchEligibilitySupportedVersionSuccess(t *t
 	validationQueryStatStatements := "SELECT extname FROM pg_extension"
 	mock.ExpectQuery(regexp.QuoteMeta(validationQueryStatStatements)).WillReturnRows(sqlmock.NewRows([]string{"extname"}).AddRow("pg_stat_statements"))
 	enabledExtensions, _ := FetchAllExtensions(conn)
-	isExtensionEnabledTest, _ := CheckBlockingSessionMetricsFetchEligibility(enabledExtensions, version)
+	isExtensionEnabledTest := CheckBlockingSessionMetricsFetchEligibility(enabledExtensions, version)
 	assert.Equal(t, isExtensionEnabledTest, true)
 	assert.NoError(t, mock.ExpectationsWereMet())
 }
@@ -35,7 +35,7 @@ func TestCheckBlockingSessionMetricsFetchEligibilitySupportedVersionFail(t *test
 	validationQueryStatStatements := "SELECT extname FROM pg_extension"
 	mock.ExpectQuery(regexp.QuoteMeta(validationQueryStatStatements)).WillReturnRows(sqlmock.NewRows([]string{"extname"}).AddRow("pg_stat_statements"))
 	enabledExtensions, _ := FetchAllExtensions(conn)
-	isExtensionEnabledTest, _ := CheckBlockingSessionMetricsFetchEligibility(enabledExtensions, version)
+	isExtensionEnabledTest := CheckBlockingSessionMetricsFetchEligibility(enabledExtensions, version)
 	assert.Equal(t, isExtensionEnabledTest, true)
 	assert.NoError(t, mock.ExpectationsWereMet())
 }
@@ -45,7 +45,7 @@ func TestIndividualQueryMetricsFetchEligibilitySupportedVersionSuccess(t *testin
 	validationQueryStatStatements := "SELECT extname FROM pg_extension"
 	mock.ExpectQuery(regexp.QuoteMeta(validationQueryStatStatements)).WillReturnRows(sqlmock.NewRows([]string{"extname"}).AddRow("pg_stat_monitor"))
 	enabledExtensions, _ := FetchAllExtensions(conn)
-	isExtensionEnabledTest, _ := CheckIndividualQueryMetricsFetchEligibility(enabledExtensions)
+	isExtensionEnabledTest := CheckIndividualQueryMetricsFetchEligibility(enabledExtensions)
 	assert.Equal(t, isExtensionEnabledTest, true)
 	assert.NoError(t, mock.ExpectationsWereMet())
 }
@@ -55,30 +55,56 @@ func TestIndividualQueryMetricsFetchEligibilitySupportedVersionFail(t *testing.T
 	validationQueryStatStatements := "SELECT extname FROM pg_extension"
 	mock.ExpectQuery(regexp.QuoteMeta(validationQueryStatStatements)).WillReturnRows(sqlmock.NewRows([]string{"extname"}))
 	enabledExtensions, _ := FetchAllExtensions(conn)
-	isExtensionEnabledTest, _ := CheckIndividualQueryMetricsFetchEligibility(enabledExtensions)
+	isExtensionEnabledTest := CheckIndividualQueryMetricsFetchEligibility(enabledExtensions)
 	assert.Equal(t, isExtensionEnabledTest, false)
 	assert.NoError(t, mock.ExpectationsWereMet())
 }
 
 func TestCheckWaitEventMetricsFetchEligibility(t *testing.T) {
-	validationQuery := "SELECT extname FROM pg_extension"
 	testCases := []struct {
-		waitExt  string
-		statExt  string
-		expected bool
+		name              string
+		enabledExtensions map[string]bool
+		expected          bool
 	}{
-		{"pg_wait_sampling", "pg_stat_statements", true}, // Success
-		{"pg_wait_sampling", "", false},                  // Fail V1
-		{"", "pg_stat_statements", false},                // Fail V2
+		{
+			name: "Both pg_wait_sampling and pg_stat_statements enabled",
+			enabledExtensions: map[string]bool{
+				"pg_wait_sampling":   true,
+				"pg_stat_statements": true,
+			},
+			expected: true,
+		},
+		{
+			name: "Only pg_stat_statements enabled",
+			enabledExtensions: map[string]bool{
+				"pg_wait_sampling":   false,
+				"pg_stat_statements": true,
+			},
+			expected: true,
+		},
+		{
+			name: "Neither pg_wait_sampling nor pg_stat_statements enabled",
+			enabledExtensions: map[string]bool{
+				"pg_wait_sampling":   false,
+				"pg_stat_statements": false,
+			},
+			expected: false,
+		},
+		{
+			name: "Only pg_wait_sampling enabled",
+			enabledExtensions: map[string]bool{
+				"pg_wait_sampling":   true,
+				"pg_stat_statements": false,
+			},
+			expected: false,
+		},
 	}
 
-	conn, mock := connection.CreateMockSQL(t)
 	for _, tc := range testCases {
-		mock.ExpectQuery(regexp.QuoteMeta(validationQuery)).WillReturnRows(sqlmock.NewRows([]string{"extname"}).AddRow(tc.waitExt).AddRow(tc.statExt))
-		enabledExtensions, _ := FetchAllExtensions(conn)
-		isExtensionEnabledTest, _ := CheckWaitEventMetricsFetchEligibility(enabledExtensions)
-		assert.Equal(t, isExtensionEnabledTest, tc.expected)
-		assert.NoError(t, mock.ExpectationsWereMet())
+		t.Run(tc.name, func(t *testing.T) {
+			result := CheckWaitEventMetricsFetchEligibility(tc.enabledExtensions)
+			assert.Equal(t, tc.expected, result)
+		})
 	}
 }
 
@@ -87,7 +113,7 @@ func TestCheckSlowQueryMetricsFetchEligibilitySupportedVersionSuccess(t *testing
 	validationQueryStatStatements := "SELECT extname FROM pg_extension"
 	mock.ExpectQuery(regexp.QuoteMeta(validationQueryStatStatements)).WillReturnRows(sqlmock.NewRows([]string{"extname"}).AddRow("pg_stat_statements"))
 	enabledExtensions, _ := FetchAllExtensions(conn)
-	isExtensionEnabledTest, _ := CheckSlowQueryMetricsFetchEligibility(enabledExtensions)
+	isExtensionEnabledTest := CheckSlowQueryMetricsFetchEligibility(enabledExtensions)
 	assert.Equal(t, isExtensionEnabledTest, true)
 	assert.NoError(t, mock.ExpectationsWereMet())
 }
@@ -97,7 +123,7 @@ func TestCheckSlowQueryMetricsFetchEligibilitySupportedVersionFail(t *testing.T)
 	validationQueryStatStatements := "SELECT extname FROM pg_extension"
 	mock.ExpectQuery(regexp.QuoteMeta(validationQueryStatStatements)).WillReturnRows(sqlmock.NewRows([]string{"extname"}))
 	enabledExtensions, _ := FetchAllExtensions(conn)
-	isExtensionEnabledTest, _ := CheckSlowQueryMetricsFetchEligibility(enabledExtensions)
+	isExtensionEnabledTest := CheckSlowQueryMetricsFetchEligibility(enabledExtensions)
 	assert.Equal(t, isExtensionEnabledTest, false)
 	assert.NoError(t, mock.ExpectationsWereMet())
 }